rav1e: fix wrong prefix in pkgconfig
Summary:
Pkgconfig is generated during execution of `cargo cbuild`, and it relies on the prefix that's provided in `--prefix` argument. If argument isn't provided, it defaults to `/usr/local`, which breaks compilation of software that gets information from pkgconfig.
Depends on D10802.

Test Plan: Successfully compiled software that uses librav1e and gets information from pkgconfig.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D10831